Rename: domain 패키지의 하위패키지 생성
