+++ src/main/java/christmas/domain/menu/MenuCategory.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.menu;
 
 public enum MenuCategory {
     APPETIZER,
