+++ src/main/java/christmas/domain/orders/OrderItem.java
@@ -1,5 +1,7 @@
-package christmas.domain;
+package christmas.domain.orders;
 
+import christmas.domain.menu.MenuCategory;
+import christmas.domain.menu.MenuItem;
 import christmas.validationUtils.OrderItemValidationUtils;
 
 public class OrderItem {
