+++ src/main/java/christmas/service/EventService.java
@@ -1,7 +1,7 @@
 package christmas.service;
 
-import christmas.domain.Orders;
-import christmas.domain.VisitingDate;
+import christmas.domain.orders.Orders;
+import christmas.domain.visitingDate.VisitingDate;
 
 public class EventService {
     private final VisitingDate date;
