+++ src/test/java/christmas/domain/OrderItemTest.java
@@ -1,5 +1,6 @@
 package christmas.domain;
 
+import christmas.domain.orders.OrderItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
