+++ src/test/java/christmas/domain/OrdersTest.java
@@ -1,5 +1,7 @@
 package christmas.domain;
 
+import christmas.domain.orders.OrderItem;
+import christmas.domain.orders.Orders;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
