+++ src/test/java/christmas/domain/VisitingDateTest.java
@@ -1,5 +1,6 @@
 package christmas.domain;
 
+import christmas.domain.visitingDate.VisitingDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
